pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cse-stock-pipeline"
version = "0.1.0"
description = "CSE Stock Market Data Pipeline — automated daily comp sheet ingestion"
requires-python = ">=3.11"
dependencies = [
    # Gmail API
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.120.0",
    # HTTP
    "requests>=2.31.0",
    # Microsoft OAuth (OneDrive/SharePoint)
    "msal>=1.26.0",
    # Browser automation (SharePoint downloads)
    "playwright>=1.40.0",
    # Config
    "python-dotenv>=1.0.0",
    # PDF (Week 2)
    "pdfplumber>=0.10.3",
    # Data (Week 2+)
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    # Database (Week 3)
    "sqlalchemy>=2.0.28",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    # API (Week 4)
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    # Scheduling (Week 3)
    "apscheduler>=3.10.4",
    # LLM (Week 2)
    "anthropic>=0.21.0",
    "openai>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.2.0",
    "flake8>=7.0.0",
    "flake8-docstrings>=1.7.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "locust>=2.24.0",
    "httpx>=0.27.0",
]

[project.scripts]
cse-pipeline = "src.pipeline:main"
cse-backlog = "src.backlog_processor:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ─── pytest ───────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80 -v"
markers = [
    "integration: marks tests that call external APIs (deselect with -m 'not integration')",
    "unit: marks fast unit tests with no external calls",
]

# ─── black ────────────────────────────────────────────────────────────────────
[tool.black]
line-length = 88
target-version = ["py311"]

# ─── flake8 (via setup.cfg-style) ────────────────────────────────────────────
[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "D100", "D101", "D102", "D103", "D104"]
per-file-ignores = ["tests/*:D"]

# ─── mypy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["tests/", "infrastructure/"]
